+++ app/controllers/pay/payments_controller.rb
@@ -4,8 +4,9 @@ module Pay
 
     before_action :set_redirect_to
 
+    # Intents on a Stripe Connect account are linked with ?stripe_account=acct_123
     def show
-      @payment = Payment.from_id(params[:id])
+      @payment = Payment.from_id(params[:id], stripe_account: params[:stripe_account].presence)
     rescue ::Stripe::StripeError => e
       redirect_to root_path, alert: e.message
     end