+++ test/pay/stripe/webhooks/payment_action_required_test.rb
@@ -62,4 +62,16 @@ class Pay::Stripe::Webhooks::PaymentActionRequiredTest < ActiveSupport::TestCase
       Pay::Stripe::Webhooks::PaymentActionRequired.new.call(event)
     end
   end
+
+  test "passes the connected account to the email" do
+    event = stripe_event("invoice.payment_action_required", account: "acct_123")
+    Pay::Stripe::Subscription.sync event.data.object.subscription, object: fake_stripe_subscription(id: event.data.object.subscription, customer: event.data.object.customer, status: :past_due)
+    ::Stripe::InvoicePayment.stubs(:list).returns(::Stripe::ListObject.construct_from(object: "list", data: [{id: "inpay_1234", object: "invoice_payment", status: "open", payment: {type: "payment_intent", payment_intent: "pi_1234"}}]))
+
+    mailer = mock
+    mailer.expects(:payment_action_required).returns(stub(deliver_later: true))
+    Pay.mailer.expects(:with).with(has_entries(payment_intent_id: "pi_1234", stripe_account: "acct_123")).returns(mailer)
+
+    Pay::Stripe::Webhooks::PaymentActionRequired.new.call(event)
+  end
 end